requirements.txt
# Project dependencies
fastmcp>=0.4.1
httpx>=0.28.1
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
requests==2.31.0
urllib3==2.1.0
python-dotenv==1.0.0
pydantic>=2.0.0


